blob loading: remove arbitrary bounds check
SF Bug: 3562547

 The blob loading/unloading routines have an arbitrary bound check that
was hosing our ability to pack data in the RPC code when it got too big.
The checks were put in initially because all the TPM driver used a 4K
max buffer to send blobs to the TPM chip and the loading functions
weren't used for anything other than that. Now TPM drivers with even
smaller buffers exist and the routines are used for RPC data packing, so
the bounds check makes no sense.

Signed-off-by: Kent Yoder <key@linux.vnet.ibm.com>