tcsd: Only read 4K at a time
TCSD blindly allocated the amount of data given to it in the first 4
bytes of the TSP's packet, trivially allowing a TSP to force a 2GB
allocation. Rework the TCSD's receive routine to stage allocations 4K at
a time to require the TSP to send all the data before the allocation
completes.

Signed-off-by: Kent Yoder <key@linux.vnet.ibm.com>