BSD Re-license
All authors of the TrouSerS project have given explicit permission to
re-license their contributions under the 3 clause BSD License in order for
the project to gain further adoption and encourage GPL licensed projects
to be able to link to the TrouSerS project.

To reflect this change the LICENSE file is being replaced.

Signed-off-by: Joel Schopp <jschopp@linux.vnet.ibm.com>