Add support for running tcsd as non TSS:TSS user+group.
When developing with tcsd it is currently required to create the user/group
TSS:TSS since tcsd will switch to it and check that config and state files
are owned by it. This can be troublesome in development environments where
you want to run tcsd as a user and not mess with global system settings.

Add configure switch --disable-usercheck for disable the TSS:TSS need of tcsd:
- No tcsd.conf and /var/lib/tpm owner-checking
- No switching to user by tcsd
- No useradd/groupadd and userdel/groupdel
- No chowns and chmods
- Warning at start of tcsd about this behaviour

Signed-off-by: Andreas Fuchs <andreas.fuchs@sit.fraunhofer.de>
Acked-by: Joel Schopp <jschopp@linux.vnet.ibm.com>
Tested-by: Richard Maciel <rmaciel@linux.vnet.ibm.com>