Don't allow TrouSerS to wrap asymmetric key > 2048 bits
Previously the tspi layer accepted external asymmetric keys of any
size to be wrapped by the TPM, however it allocated buffers capable
of only holding private keys with at most 2048 bits.

A test was added to prevent this behavior, since the TPM 1.1b and
1.2 limits that size, according to the TPM_STORE_ASYMKEY structure.

Also, comments and defines were added to make the life of future
maintainers easier when debugging that function.

Signed-off-by: Richard Maciel <rmaciel@linux.vnet.ibm.com>