Fixed memory leak in ima_get_entry and fixed memset boundaries
Related to coverity CID 10311.

In some error cases the memory allocated wasn't being properly released,
so I grouped all the release in the end of the function and make error
cases point to the label there.

Also fixed a memset call which used a size 1-byte less than the
correct amount.

Signed-off-by: Richard Maciel <rmaciel@linux.vnet.ibm.com>