Fixed memory leak in function Transport_TerminateHandle
Related to coverity CID 10307.

Pointer handles got memory allocated for it, but that memory is never
freed at the end of the function.
Note that since obj_context_transport_execute can also allocate memory
(through the same parameter handles), it's necessary to keep track of
both the address which goes in and the address which cames out.

Signed-off-by: Richard Maciel <rmaciel@linux.vnet.ibm.com>